Fixed astropy.wcs import
The WCS package in astropy is at astropy.wcs not astropy.pywcs. Since astropy.wcs is guaranteed to be a very recent version of pywcs, I thought it would make more sense to try that first. For example, if someone has an old installation of pywcs and astropy, astropy should be preferred (to have the 'sub' attribute).